disable DVB support to fix building against linux 5.14 headers
This change fixes building MPV on Fedora 35+ (and will probably have
to be backported to Fedora 34 as well, once the kernel is updated
to 5.14 there too).

The /usr/include/linux/dvb/video.h header file has been (re)moved,
with the DVB drivers having been moved to the staging tree.